Encode domain name label value
Since we figured out that the in `CFApp` and `CFDomain` resources the
display name can contain spaces and be longer than 63 characters, but
the kubernetes lables values cannot be longer than 63 characters and
containing spaces, we decided to encode with Sha224 the `Spec.Name`
string in the label selector. In this way when the 'List' to the resource is called
it encodes the displayName string from the message in the repository and
compares it with the encode value in label selector.
The encoded label values are set in defalter webhooks.

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>